refactor: extract CurrencyFormatted
